tested printf + gnl changes

+++ printf/printf_libft/libft.h
@@ -20,7 +20,7 @@
 int		ft_atoi(const char *str);
 void	ft_bzero(void *s, size_t n);
 void	*ft_calloc(size_t count, size_t size);
-char	*ft_hexstr(unsigned long n, int *k);
+char	*ft_hexstr(unsigned int n, int *k);
 int		ft_isalnum(int c);
 int		ft_isalpha(int c);
 int		ft_isascii(int c);
